# openapi.yaml
openapi: 3.0.0

info:
  title: Go template app
  version: 1.0.0
security:
  - BearerAuth: []
paths:
  /hello:
    get:
      tags:
        - hello-world
      summary: "Helloworld is a default Get method used as example"
      operationId: "HelloWorld"
      responses:
        "200":
          $ref: "#/components/responses/SuccessResponse"
        "400":
          description: Bad Request
          $ref: "#/components/responses/BadRequestResponse"
        "404":
          description: Not Found
          $ref: "#/components/responses/NotFoundResponse"
        "5XX":
          description: Unexpected error
          $ref: "#/components/responses/InternalErrorResponse"
components:
  responses:
    SuccessResponse:
      description: Success Rest Response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Success"
          example:   # Sample object
              success: true
              data: 
                hello: World
    BadRequestResponse:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:   # Sample object
              success: false
              data: 
                error: Bad request
    NotFoundResponse:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:   # Sample object
              success: false
              data: 
                error: not found
    InternalErrorResponse:
      description: Fail Rest Response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:   # Sample object
              success: false
              data: 
                error: Internal Server Error
  schemas:
    Success:
      type: object
      required:
        - "success"
        - "data"
      properties:
        success:
          type: boolean
        data:
          type: object
    Error:
      type: object
      required:
        - "success"
        - "data"
      properties:
        success:
          type: boolean
        data:
          type: object
          required:
            - "success"
          properties:
            error:
              type: string